refactor(control-plane): split environment-secrets routes file
Follow-up cleanups on PR #915:

- move the four environment-secrets handlers (list/set/delete/import) plus the
  requireSecretsConfig guard into routes/environment-secrets.ts; environments.ts
  drops from 517 to 264 lines and holds CRUD only. Both route arrays are
  registered in the router.
- drop the dead `export type { SecretMetadata }` re-export from
  environment-secrets.ts (nothing consumes it; global-secrets.ts already uses
  the import-only form). repo-secrets.ts carries the same dead line — left for
  the scoped-secrets unification refactor.
- remove decorative section-separator comments from db/environments.ts.
